make_macs_xls.py: fix bug from typo in previous commit.
