+++ bin/make_macs2_xls.py
@@ -1,0 +1,8 @@
+#!/usr/bin/env python
+#
+#     make_macs2_xls.py: Convert MACS output file to XLS(X) spreadsheet
+#     Copyright (C) University of Manchester 2013-2021 Peter Briggs, Ian Donaldson
+#
+from bcftbx.cli.make_macs2_xls import main
+if __name__ == "__main__":
+     main()
